feat: 新增工单处理流程与确认报告功能
1. 新增工单处理状态：已处理、待反馈调整，扩展工单状态枚举与前端样式
2. 新增工单处理进程表与相关服务，支持处理记录与回复树状展示
3. 扩展附件类型，新增确认报告类型，支持PDF/图片上传校验
4. 新增确认报告导出Word与打印页面，支持快速生成标准化文档
5. 优化归档节点排序功能，支持小数序号插入调整
6. 扩展数据库表结构与迁移脚本，适配新增功能

+++ src/main/java/com/qs/dto/ArchiveNodeStageRequest.java
@@ -3,7 +3,8 @@ package com.qs.dto;
 public class ArchiveNodeStageRequest {
 
     private String name;
-    private Integer sortOrder;
+    /** 支持小数：如 1.55 表示插入到 1 与 2 之间，服务端进位并为后续序号腾位 */
+    private Double sortOrder;
     private String colorKey;
 
     public String getName() {
@@ -14,11 +15,11 @@ public class ArchiveNodeStageRequest {
         this.name = name;
     }
 
-    public Integer getSortOrder() {
+    public Double getSortOrder() {
         return sortOrder;
     }
 
-    public void setSortOrder(Integer sortOrder) {
+    public void setSortOrder(Double sortOrder) {
         this.sortOrder = sortOrder;
     }
